+++ HISTORY.rst
@@ -1,6 +1,12 @@
 History
 =======
 
+0.5.3 (TBD)
+-----------
+
+* Bugfix: exceptions no longer swallowed when accessing from context a
+  property that raises an exception (issue #110).
+
 0.5.2 (2012-05-03)
 ------------------